feat: upload backup to all defined rclone remotes
This change also moves the backup path from a subdirectory named after the bot container (created during backup if previously non-existent) to the root of the rclone remote.

The command to prune old backups is also updated to target files that look like a labbot backup, making it a little safer as it's no longer guaranteed to be operating in a dedicated directory.